// packages/atomic-library/src/molecules/MlBanner/styledComponents/types.ts
import { ThemeType } from "../../../theme";

export interface StyledPressableProps {
  theme: ThemeType;
  disabled?: boolean;
  width: number;
}

export interface StyledImageWrapperProps {
  theme: ThemeType;
  width: number;
  height: number;
}

export interface StyledDetailContainerProps {
  theme: ThemeType;
  width: number;
}